Update to use JPA finished

+++ src/main/java/fastvagas/data/repository/CityRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CityRepository extends JpaRepository<City, Integer> {
+public interface CityRepository extends JpaRepository<City, Long> {
 
-    List<City> findAllByStateId(Integer stateId);
+    List<City> findAllByStateId(Long stateId);
 }
