+++ src/main/java/fastvagas/data/repository/StateRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface StateRepository extends JpaRepository<State, Integer> {
+public interface StateRepository extends JpaRepository<State, Long> {
 
     Optional<State> findByAcronym(String acronym);
 }
